requirements.txt
binance-connector
binance-futures-connector
PySide6